refactor(acp): extract NdjsonRpcTransport from acp-lifecycle-probe.ts (#4235 step 2)
